chemical-rag-system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
rdkit==2026.03.1
faiss-cpu==1.13.2
numpy==2.0.2
pillow==10.1.0
pubchempy==1.0.5
pydantic==2.5.0
huggingface_hub==0.21.4
